# api/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "UI起点・初心者向け OSS DAST（Nuclei ラッパー）の API。スキャン受付・サイト管理・ドメイン所有確認・認証情報の暗号化保管・ダッシュボード/レポート用データ提供。\n認証情報の生値はレスポンスに一切含めない（マスク）。スキャン実行は worker プロセスに分離（ADR-0001）。",
        "title": "goodast API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/scans/{id}": {
            "get": {
                "description": "スキャンの status とサマリ（done は score/band/counts、未完了は summary=null）を返す。進捗ポーリング兼用。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "スキャン状態を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.ScanState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scans/{id}/findings": {
            "get": {
                "description": "findings を重大度の重い順（Critical→Info）で返す。クリーン（0件）は空配列、scan 不在は 404。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "スキャンの findings 明細を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.findingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/sites/{id}/credentials": {
            "get": {
                "description": "auth_mode と configured を返す（生値は含めない）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "認証情報の状態を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.credentialStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "持ち込みセッション（ヘッダ）をアプリ層で暗号化保存（ADR-0003）。生値はレスポンスに含めない。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "認証情報を設定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "認証ヘッダ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.setCredentialsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.credentialStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "credentials"
                ],
                "summary": "認証情報を削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sites/{id}/dashboard": {
            "get": {
                "description": "最新スコア＋前回差分＋スコア時系列（done スキャンのみ）。スキャン無し/未知サイトは 200＋latest=null・history=[]。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "サイトのダッシュボード集計を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.DashboardData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sites/{id}/scans": {
            "get": {
                "description": "サイトの過去スキャンを新しい順で返す（全 status）。各エントリは GET /scans/{id} と同形。未知サイトは 200＋scans=[]。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "サイトの診断履歴を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.scanHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ymd38_goodast_api_internal_report.Band": {
            "type": "string",
            "enum": [
                "good",
                "caution",
                "danger",
                "crisis"
            ],
            "x-enum-comments": {
                "BandCaution": "要注意（60〜79）→ --color-warning",
                "BandCrisis": "危機（0〜39）→ --color-m-red + opacity 強調",
                "BandDanger": "危険（40〜59）→ --color-m-red",
                "BandGood": "良好（80〜100）→ --color-success"
            },
            "x-enum-descriptions": [
                "良好（80〜100）→ --color-success",
                "要注意（60〜79）→ --color-warning",
                "危険（40〜59）→ --color-m-red",
                "危機（0〜39）→ --color-m-red + opacity 強調"
            ],
            "x-enum-varnames": [
                "BandGood",
                "BandCaution",
                "BandDanger",
                "BandCrisis"
            ]
        },
        "github_com_ymd38_goodast_api_internal_report.DashboardData": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.HistoryEntry"
                    }
                },
                "latest": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.LatestState"
                }
            }
        },
        "github_com_ymd38_goodast_api_internal_report.Finding": {
            "type": "object",
            "properties": {
                "cwe": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "remediation": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_ymd38_goodast_api_internal_report.HistoryEntry": {
            "type": "object",
            "properties": {
                "band": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.Band"
                },
                "counts": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.SeverityCounts"
                },
                "date": {
                    "type": "string"
                },
                "scan_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "github_com_ymd38_goodast_api_internal_report.LatestState": {
            "type": "object",
            "properties": {
                "band": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.Band"
                },
                "counts": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.SeverityCounts"
                },
                "date": {
                    "type": "string"
                },
                "delta": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "scan_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "github_com_ymd38_goodast_api_internal_report.ScanState": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "engine_version": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "site_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.ScanSummary"
                }
            }
        },
        "github_com_ymd38_goodast_api_internal_report.ScanSummary": {
            "type": "object",
            "properties": {
                "band": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.Band"
                },
                "counts": {
                    "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.SeverityCounts"
                },
                "label": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "github_com_ymd38_goodast_api_internal_report.SeverityCounts": {
            "type": "object",
            "properties": {
                "critical": {
                    "type": "integer"
                },
                "high": {
                    "type": "integer"
                },
                "info": {
                    "type": "integer"
                },
                "low": {
                    "type": "integer"
                },
                "medium": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid site id"
                }
            }
        },
        "internal_handler.credentialStatusResponse": {
            "type": "object",
            "properties": {
                "auth_mode": {
                    "type": "string"
                },
                "configured": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                }
            }
        },
        "internal_handler.findingsResponse": {
            "type": "object",
            "properties": {
                "findings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.Finding"
                    }
                },
                "scan_id": {
                    "type": "string"
                }
            }
        },
        "internal_handler.headerInput": {
            "type": "object",
            "required": [
                "name",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.scanHistoryResponse": {
            "type": "object",
            "properties": {
                "scans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ymd38_goodast_api_internal_report.ScanState"
                    }
                },
                "site_id": {
                    "type": "string"
                }
            }
        },
        "internal_handler.setCredentialsRequest": {
            "type": "object",
            "required": [
                "headers"
            ],
            "properties": {
                "headers": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/internal_handler.headerInput"
                    }
                }
            }
        },
...
    }
}